refactor(tui): move state test module into state/tests.rs
The inline #[cfg(test)] module was ~2.7k lines — more than the rest of
state.rs combined. Move it to its own state/tests.rs file (mod tests;),
leaving state/mod.rs at ~370 lines: AppState, its accessors, the public
reduce() wrapper, and shared consts/helpers. `super` still resolves to the
state module, so every use path is unchanged. No behavior change.
